third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/cross-origin-unsupported-url-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:1291a/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:1291a/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load ftp://127.0.0.1/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load ftp://127.0.0.1/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load localhost:8080/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load localhost:8080/. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load tel:1234. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: XMLHttpRequest cannot load tel:1234. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
Test cross-origin XHRs to CORS-unsupported protocol schemes in the URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS errorEvent.type is "error"
PASS xhr.send() did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
